Task3 Done. Start Task 4

+++ wts-backend/src/main/java/com/wts/njp/model/User.java
@@ -16,7 +16,7 @@ public class User {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Integer id;
+	private Long id;
 	
 	@Column(name="name")
 	private String name;
@@ -31,17 +31,17 @@ public class User {
 		
 	}
 	
-	public User(Integer id, String name, Integer age, String email) {
+	public User(Long id, String name, Integer age, String email) {
 		this.id = id;
 		this.name = name;
 		this.age = age;
 		this.email = email;
 	}
 	
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
-	public void setId(Integer id) {
+	public void setId(Long id) {
 		this.id = id;
 	}
 	public String getName() {
